Refactor: extract EquipmentExporter for tabular data generation
- Moved data(), _generate_data(), _extract_path_row(), and _merge_path_data() to EquipmentExporter.
- Updated Equipment to delegate to EquipmentExporter.
- Improved modularity and maintainability of the equipment module.
